pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cql-track"
version = "0.1.0"
description = "Schema migration tool for Apache Cassandra"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "cassandra-driver>=3.28",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
cqltrack = "cqltrack.cli:main"

[tool.setuptools.packages.find]
include = ["cqltrack*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
